requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
mysql-connector-python==8.0.33
Werkzeug==2.3.7